feat(vm): Gate unix-like eh_personality implementation behind a cfg check
...as libunwind symbols are not found in every target (e.g. win-msvc)
